Move interface to `Http.Client`.
