+++ framework/src/Volo.Abp.Http.Client/Volo/Abp/Http/Client/ClientProxying/IObjectToFormData.cs
@@ -2,7 +2,7 @@
 using System.Net.Http;
 using System.Threading.Tasks;
 
-namespace Volo.Abp.Http
+namespace Volo.Abp.Http.Client.ClientProxying
 {
     public interface IObjectToFormData<in TValue>
     {
