+++ framework/src/Volo.Abp.Http.Client/Volo/Abp/Http/Client/ClientProxying/IObjectToQueryString.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace Volo.Abp.Http
+namespace Volo.Abp.Http.Client.ClientProxying
 {
     public interface IObjectToQueryString<in TValue>
     {
